// packages/discord/src/operations/updateWebhookMessage.ts
export const UpdateWebhookMessageInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    webhook_id: Schema.String.pipe(T.PathParam()),
    webhook_token: Schema.String.pipe(T.PathParam()),
    message_id: Schema.String.pipe(T.PathParam()),
    thread_id: Schema.optional(Schema.String),
    with_components: Schema.optional(Schema.Boolean),
    content: Schema.optional(Schema.NullOr(Schema.String)),
    embeds: Schema.optional(
      Schema.NullOr(
        Schema.Array(
          Schema.Struct({
            type: Schema.optional(Schema.NullOr(Schema.String)),
            url: Schema.optional(Schema.NullOr(Schema.String)),
            title: Schema.optional(Schema.NullOr(Schema.String)),
            color: Schema.optional(Schema.NullOr(Schema.Number)),
            timestamp: Schema.optional(Schema.NullOr(Schema.String)),
            description: Schema.optional(Schema.NullOr(Schema.String)),
            author: Schema.optional(Schema.Unknown),
            image: Schema.optional(Schema.Unknown),
            thumbnail: Schema.optional(Schema.Unknown),
            footer: Schema.optional(Schema.Unknown),
            fields: Schema.optional(
              Schema.NullOr(
                Schema.Array(
                  Schema.Struct({
                    name: Schema.String,
                    value: Schema.String,
                    inline: Schema.optional(Schema.NullOr(Schema.Boolean)),
                  }),
                ),
              ),
            ),
            provider: Schema.optional(Schema.Unknown),
            video: Schema.optional(Schema.Unknown),
          }),
        ),
      ),
    ),
    allowed_mentions: Schema.optional(Schema.Unknown),
    components: Schema.optional(Schema.NullOr(Schema.Array(Schema.Unknown))),
    attachments: Schema.optional(
      Schema.NullOr(
        Schema.Array(
          Schema.Struct({
            id: Schema.String,
            filename: Schema.optional(Schema.NullOr(Schema.String)),
            description: Schema.optional(Schema.NullOr(Schema.String)),
            duration_secs: Schema.optional(Schema.NullOr(Schema.Number)),
            waveform: Schema.optional(Schema.NullOr(Schema.String)),
            title: Schema.optional(Schema.NullOr(Schema.String)),
            is_remix: Schema.optional(Schema.NullOr(Schema.Boolean)),
          }),
        ),
      ),
    ),
    poll: Schema.optional(Schema.Unknown),
    flags: Schema.optional(Schema.NullOr(Schema.Number)),
  }).pipe(
    T.Http({
      method: "PATCH",
      path: "/webhooks/{webhook_id}/{webhook_token}/messages/{message_id}",
    }),
  );
export type UpdateWebhookMessageInput = typeof UpdateWebhookMessageInput.Type;
// ...
export type UpdateWebhookMessageOutput = typeof UpdateWebhookMessageOutput.Type;

// The operation
export const updateWebhookMessage = /*@__PURE__*/ /*#__PURE__*/ API.make(
  () => ({
    inputSchema: UpdateWebhookMessageInput,
    outputSchema: UpdateWebhookMessageOutput,
    errors: [BadRequest, Forbidden, NotFound] as const,
  }),
);